+++ Reflection/test/utopia/reflection/test/swing/StackContentManagementTest.scala
@@ -1,18 +1,19 @@
 package utopia.reflection.test.swing
 
+import utopia.firmament.controller.data.ContainerContentDisplayer
+import utopia.firmament.model.enumeration.WindowResizePolicy.UserAndProgram
+import utopia.firmament.model.stack.LengthExtensions._
 import utopia.flow.async.process.Loop
-import utopia.flow.time.TimeExtensions._
 import utopia.flow.collection.CollectionExtensions._
+import utopia.flow.time.TimeExtensions._
+import utopia.flow.view.mutable.eventful.PointerWithEvents
+import utopia.paradigm.color.ColorRole.Gray
 import utopia.paradigm.generic.ParadigmDataType
 import utopia.reflection.component.swing.label.ItemLabel
 import utopia.reflection.container.swing.layout.multi.Stack
 import utopia.reflection.container.swing.window.Frame
-import utopia.firmament.model.enumeration.WindowResizePolicy.User
-import utopia.reflection.controller.data.ContainerContentManager
 import utopia.reflection.test.TestContext
 import utopia.reflection.util.SingleFrameSetup
-import utopia.firmament.model.stack.LengthExtensions._
-import utopia.paradigm.color.ColorRole.Gray
 
 /**
   * Tests stack content management
@@ -29,8 +30,9 @@ object StackContentManagementTest extends App
 	val stack = Stack.column[ItemLabel[Int]]()
 	
 	val background = colorScheme(Gray)
+	val contentPointer = new PointerWithEvents(Vector(1, 4, 6))
 	val manager = baseContext.against(background).forTextComponents.use { implicit txc =>
-		ContainerContentManager.forStatelessItems[Int, ItemLabel[Int]](stack, Vector(1, 4, 6)) { i =>
+		ContainerContentDisplayer.forStatelessItems(stack, contentPointer) { i: Int =>
 			println(s"Creating a new label ($i)")
 			val label = ItemLabel.contextual(i)
 			label.contentPointer.addContinuousListener { e => println(s"Label content changed: $e") }
@@ -41,18 +43,21 @@ object StackContentManagementTest extends App
 	val content = stack.framed(64.any x 16.any, background)
 	
 	// Starts test
-	val frame = Frame.windowed(content, "Component to Image Test", User)
+	val frame = Frame.windowed(content, "Component to Image Test", UserAndProgram)
 	frame.setToCloseOnEsc()
 	new SingleFrameSetup(actorHandler, frame).start()
 	
 	println("Starting the loop")
-	var nextNumber = 8
-	Loop.regularly(1.seconds) {
-			println(s"Updating content by adding $nextNumber")
-			val index = (math.random() * manager.content.size).toInt
-			println(s"\tInserts $nextNumber to index $index")
-			manager.content = manager.content.inserted(nextNumber, index)
-			nextNumber += 2
-			println("\tNumbers updated")
+	val numberIter = Iterator.iterate(8) { _ + 2 }
+	Loop.regularly(1.5.seconds) {
+		contentPointer.update { content =>
+			if (content.size > 12)
+				content.take(2)
+			else {
+				val index = (math.random() * content.size).toInt
+				content.take(index) ++ numberIter.takeNext(if (math.random() < 0.5) 1 else 2) ++ content.drop(index)
+			}
+		}
+		println("\tNumbers updated")
 	}
 }